// polytrader-console/src/lib/api/models/health-gate-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Polytrader Control API
 * Control plane API for Polytrader platform
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Individual health gate status.
 */
export interface HealthGateStatus {
    /**
     * Gate status: ok, degraded, or down
     */
    'status': HealthGateStatusStatusEnum;
    'message'?: string | null;
}

export const HealthGateStatusStatusEnum = {
    Ok: 'ok',
    Degraded: 'degraded',
    Down: 'down'
} as const;

export type HealthGateStatusStatusEnum = typeof HealthGateStatusStatusEnum[keyof typeof HealthGateStatusStatusEnum];